requirements.txt
yfinance==0.2.43
transformers==4.45.2
torch==2.4.1
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
scikit-learn==1.5.2
tqdm==4.66.5
pyyaml==6.0.2


